+++ admin/class-pirateforms-admin.php
@@ -63,15 +63,16 @@ class PirateForms_Admin {
 	public function enqueue_styles_and_scripts() {
 		global $pagenow;
 
-		if ( ! empty( $pagenow ) && ( $pagenow == 'options-general.php' || $pagenow == 'admin.php' )
-			 && isset( $_GET['page'] ) && $_GET['page'] == 'pirateforms-admin'
-		) {
+		$allowed    = array( 'options-general.php', 'admin.php', 'edit.php', 'post.php' );
+
+		if ( ( ! empty( $pagenow ) && in_array( $pagenow, $allowed ) ) || ( isset( $_GET['page'] ) && $_GET['page'] == 'pirateforms-admin' ) ) {
 			wp_enqueue_style( 'pirateforms_admin_styles', PIRATEFORMS_URL . 'admin/css/wp-admin.css', array(), $this->version );
-			wp_enqueue_script( 'pirateforms_scripts_admin', PIRATEFORMS_URL . 'admin/js/scripts-admin.js', array( 'jquery' ), $this->version );
+			wp_enqueue_script( 'pirateforms_scripts_admin', PIRATEFORMS_URL . 'admin/js/scripts-admin.js', array( 'jquery', 'jquery-ui-tooltip' ), $this->version );
 			wp_localize_script(
 				'pirateforms_scripts_admin', 'cwp_top_ajaxload', array(
 					'ajaxurl' => admin_url( 'admin-ajax.php' ),
 					'nonce'   => wp_create_nonce( PIRATEFORMS_SLUG ),
+					'slug'    => PIRATEFORMS_SLUG,
 					'i10n'    => array(
 						'recaptcha' => __( 'Please specify the Site Key and Secret Key.', 'pirate-forms' ),
 					),
@@ -251,10 +252,10 @@ class PirateForms_Admin {
 		}
 
 		// the key(s) will be added to the div as class names
-		// to enable tooltip popup add 'pirate_dashicons'
+		// to enable tooltip popup add 'pirate_tooltip'
 		return apply_filters(
 			'pirate_forms_admin_controls', array(
-				'pirate_options pirate_dashicons' => array(
+				'pirate_options pirate_tooltip' => array(
 					'heading'  => __( 'Form processing options', 'pirate-forms' ),
 					'controls' => apply_filters(
 						'pirate_forms_admin_controls_for_options', array(
@@ -303,23 +304,24 @@ class PirateForms_Admin {
 									'value' => __( 'Store submissions in the database', 'pirate-forms' ),
 									'html'  => '<span class="dashicons dashicons-editor-help"></span>',
 									'desc'  => array(
-										'value' => __( 'Should the submissions be stored in the admin area? If chosen, contact form submissions will be saved under "All Entries" on the left (appears after this option is activated).', 'pirate-forms' ),
+										'value' => sprintf( '%s<br>%s', __( 'Should the submissions be stored in the admin area? If chosen, contact form submissions will be saved under "All Entries" on the left (appears after this option is activated).', 'pirate-forms' ), __( 'According to GDPR we recommend you to ask for consent in order to store user data', 'pirate-forms' ) ),
 										'class' => 'pirate_forms_option_description',
 									),
 								),
-								'default' => 'yes',
+								'default' => 'no',
 								'value'   => PirateForms_Util::get_option( 'pirateformsopt_store' ),
 								'wrap'    => array(
 									'type'  => 'div',
 									'class' => 'pirate-forms-grouped',
 								),
 								'options' => array( 'yes' => __( 'Yes', 'pirate-forms' ) ),
+								'title' => __( 'According to GDPR, we recommend you to ask for consent in order to store user data.', 'pirate-forms' ),
 							),
 							array(
 								'id'      => 'pirateformsopt_nonce',
 								'type'    => 'checkbox',
 								'label'   => array(
-									'value' => __( 'Add a nonce to the contact form:', 'pirate-forms' ),
+									'value' => __( 'Add a nonce to the contact form', 'pirate-forms' ),
 									'html'  => '<span class="dashicons dashicons-editor-help"></span>',
 									'desc'  => array(
 										'value' => __( 'Should the form use a WordPress nonce? This helps reduce spam by ensuring that the form submittor is on the site when submitting the form rather than submitting remotely. This could, however, cause problems with sites using a page caching plugin. Turn this off if you are getting complaints about forms not being able to be submitted with an error of "Nonce failed!"', 'pirate-forms' ),
@@ -352,6 +354,25 @@ class PirateForms_Admin {
 								),
 								'cols'  => 70,
 								'rows'  => 5,
+							),
+							array(
+								'id'      => 'pirateformsopt_copy_email',
+								'type'    => 'checkbox',
+								'label'   => array(
+									'value' => __( 'Add copy of mail to confirmation email', 'pirate-forms' ),
+									'html'  => '<span class="dashicons dashicons-editor-help"></span>',
+									'desc'  => array(
+										'value' => __( 'Should a copy of the email be appended to the confirmation email? Only the fields that are being displayed will be sent to the sender. Please note that this will only be appended if confirmation email text is provided above.', 'pirate-forms' ),
+										'class' => 'pirate_forms_option_description',
+									),
+								),
+								'default' => '',
+								'value'   => PirateForms_Util::get_option( 'pirateformsopt_copy_email' ),
+								'wrap'    => array(
+									'type'  => 'div',
+									'class' => 'pirate-forms-grouped',
+								),
+								'options' => array( 'yes' => __( 'Yes', 'pirate-forms' ) ),
 							),
 							array(
 								'id'      => 'pirateformsopt_thank_you_url',
@@ -395,7 +416,7 @@ class PirateForms_Admin {
 						)
 					),
 				),
-				'pirate_fields pirate_dashicons'  => array(
+				'pirate_fields pirate_tooltip'  => array(
 					'heading'  => __( 'Fields Settings', 'pirate-forms' ),
 					'controls' => apply_filters(
 						'pirate_forms_admin_controls_for_fields', array(
@@ -493,6 +514,23 @@ class PirateForms_Admin {
 									'req' => __( 'Required', 'pirate-forms' ),
 								),
 							),
+							array(
+								'id'      => 'pirateformsopt_checkbox_field',
+								'type'    => 'select',
+								'label'   => array(
+									'value' => __( 'Checkbox', 'pirate-forms' ),
+								),
+								'value'   => PirateForms_Util::get_option( 'pirateformsopt_checkbox_field' ),
+								'wrap'    => array(
+									'type'  => 'div',
+									'class' => 'pirate-forms-grouped pirateformsopt_checkbox',
+								),
+								'options' => array(
+									''    => __( 'Do not display', 'pirate-forms' ),
+									'yes' => __( 'Display but not required', 'pirate-forms' ),
+									'req' => __( 'Required', 'pirate-forms' ),
+								),
+							),
 							/* Recaptcha */
 							array(
 								'id'      => 'pirateformsopt_recaptcha_field',
@@ -546,7 +584,7 @@ class PirateForms_Admin {
 						)
 					),
 				),
-				'pirate_labels pirate_dashicons'  => array(
+				'pirate_labels pirate_tooltip'  => array(
 					'heading'  => __( 'Fields Labels', 'pirate-forms' ),
 					'controls' => apply_filters(
 						'pirate_forms_admin_controls_for_field_labels', array(
@@ -616,20 +654,37 @@ class PirateForms_Admin {
 								),
 							),
 							array(
+								'id'      => 'pirateformsopt_label_checkbox',
+								'type'    => 'wysiwyg',
+								'label'   => array(
+									'value' => __( 'Checkbox', 'pirate-forms' ),
+								),
+								'value'   => PirateForms_Util::get_option( 'pirateformsopt_label_checkbox' ),
+								'wrap'    => array(
+									'type'  => 'div',
+									'class' => 'pirate-forms-grouped',
+								),
+								'wysiwyg' => array(
+									'editor_class'  => 'pirate-forms-wysiwyg',
+									'quicktags' => false,
+									'teeny' => true,
+									'media_buttons' => false,
+								),
+							),
+							array(
 								'id'      => 'pirateformsopt_email_content',
 								'type'    => 'wysiwyg',
 								'label'   => array(
 									'value' => __( 'Email content', 'pirate-forms' ),
 									'html'  => '<br/><br/>' . esc_attr( __( 'You can use the next magic tags:', 'pirate-forms' ) ) . '<br/>' . PirateForms_Util::get_magic_tags(),
 								),
-								'default' => PirateForms_Util::get_default_email_content(),
+								'default' => PirateForms_Util::get_default_email_content( true, null, true ),
 								'value'   => PirateForms_Util::get_option( 'pirateformsopt_email_content' ),
 								'wrap'    => array(
 									'type'  => 'div',
 									'class' => 'pirate-forms-grouped',
 								),
 								'wysiwyg' => array(
-									'quicktags'     => false,
 									'editor_class'  => 'pirate-forms-wysiwyg',
 									'editor_height' => 500,
 								),
@@ -637,7 +692,7 @@ class PirateForms_Admin {
 						)
 					),
 				),
-				'pirate_alerts pirate_dashicons'  => array(
+				'pirate_alerts pirate_tooltip'  => array(
 					'heading'  => __( 'Alert Messages', 'pirate-forms' ),
 					'controls' => apply_filters(
 						'pirate_forms_admin_controls_for_alerts', array(
@@ -707,6 +762,19 @@ class PirateForms_Admin {
 								),
 							),
 							array(
+								'id'      => 'pirateformsopt_label_err_no_checkbox',
+								'type'    => 'text',
+								'label'   => array(
+									'value' => __( 'Checkbox is not checked', 'pirate-forms' ),
+								),
+								'default' => __( 'Please select the checkbox', 'pirate-forms' ),
+								'value'   => PirateForms_Util::get_option( 'pirateformsopt_label_err_no_checkbox' ),
+								'wrap'    => array(
+									'type'  => 'div',
+									'class' => 'pirate-forms-grouped',
+								),
+							),
+							array(
 								'id'      => 'pirateformsopt_label_submit',
 								'type'    => 'text',
 								'label'   => array(
@@ -727,7 +795,7 @@ class PirateForms_Admin {
 						)
 					),
 				),
-				'pirate_smtp pirate_dashicons'    => array(
+				'pirate_smtp pirate_tooltip'    => array(
 					'heading'  => __( 'SMTP Options', 'pirate-forms' ),
 					'controls' => apply_filters(
 						'pirate_forms_admin_controls_for_smtp', array(
@@ -1010,6 +1078,48 @@ class PirateForms_Admin {
 		$options['pirateformsopt_attachment_field'] = 'no';
 
 		return $options;
+	}
+
+	/**
+	 * Show admin notices.
+	 */
+	public function admin_notices() {
+		$screen = get_current_screen();
+		if ( empty( $screen ) ) {
+			return;
+		}
+		if ( ! isset( $screen->base ) ) {
+			return;
+		}
+
+		if ( ! in_array( $screen->id, array( 'toplevel_page_pirateforms-admin' ) ) ) {
+			return;
+		}
+
+		$options = PirateForms_Util::get_option();
+
+		// check if store submissions is enabled without having a checkbox.
+		if ( 'yes' !== $options['pirateformsopt_store'] ) {
+			return;
+		}
+
+		if ( empty( $options['pirateformsopt_checkbox_field'] ) && false === ( $x = get_transient( 'pirate_forms_gdpr_notice0' ) ) ) {
+			echo sprintf( '<div data-dismissible="0" class="notice notice-warning pirateforms-notice pirateforms-notice-checkbox pirateforms-notice-gdpr is-dismissible"><p><strong>%s</strong></p></div>', __( 'According to GDPR we recommend you to ask for consent in order to store user data', 'pirate-forms' ) );
+		}
+	}
+
+	/**
+	 * Generic ajax handler.
+	 */
+	public function ajax() {
+		check_ajax_referer( PIRATEFORMS_SLUG, 'security' );
+
+		switch ( $_POST['_action'] ) {
+			case 'dismiss-notice':
+				set_transient( 'pirate_forms_gdpr_notice' . $_POST['id'], 'yes' );
+				break;
+		}
+		wp_die();
 	}
 
 	/**