+++ public/class-pirateforms-public.php
@@ -41,6 +41,45 @@ class PirateForms_Public {
 	private $version;
 
 	/**
+	 * The file types allowed to be uploaded. Can take a look at https://en.support.wordpress.com/accepted-filetypes/ for inspiration.
+	 *
+	 * @access   private
+	 * @var      array $_file_types_allowed The file types allowed to be uploaded.
+	 */
+	private static $_file_types_allowed = array(
+		'3g2',
+		'3gp',
+		'avi',
+		'doc',
+		'docx',
+		'gif',
+		'jpeg',
+		'jpg',
+		'key',
+		'm4a',
+		'm4v',
+		'mov',
+		'mp3',
+		'mp4',
+		'mpg',
+		'odt',
+		'ogg',
+		'ogv',
+		'pdf',
+		'png',
+		'pps',
+		'ppsx',
+		'ppt',
+		'pptx',
+		'txt',
+		'wav',
+		'wmv',
+		'xls',
+		'xlsx',
+		'zip',
+	);
+
+	/**
 	 * Initialize the class and set its properties.
 	 *
 	 * @since    1.0.0
@@ -191,6 +230,14 @@ class PirateForms_Public {
 			/**
 			 ******  Name field */
 			if ( ! empty( $field ) && ! empty( $label ) ) {
+				$required     = $field === 'req' ? true : false;
+				$wrap_classes = array(
+					( ! empty( $subjectField ) ? 'col-xs-12 col-sm-6' : 'col-xs-12' ) . ' contact_name_wrap pirate_forms_three_inputs form_field_wrap',
+				);
+				// If this field was submitted with invalid data
+				if ( isset( $_SESSION[ $error_key ]['contact-name'] ) ) {
+					$wrap_classes[] = 'error';
+				}
 				$elements[] = array(
 					'front_end'     => true,
 					'placeholder'  => stripslashes( sanitize_text_field( $label ) ),
@@ -209,6 +256,14 @@ class PirateForms_Public {
 			/**
 			 ******  Email field */
 			if ( ! empty( $field ) && ! empty( $label ) ) {
+				$required     = $field === 'req' ? true : false;
+				$wrap_classes = array(
+					( ! empty( $subjectField ) ? 'col-xs-12 col-sm-6' : 'col-xs-12' ) . ' contact_email_wrap pirate_forms_three_inputs form_field_wrap',
+				);
+				// If this field was submitted with invalid data
+				if ( isset( $_SESSION[ $error_key ]['contact-email'] ) ) {
+					$wrap_classes[] = 'error';
+				}
 				$elements[] = array(
 					'front_end'     => true,
 					'placeholder'  => stripslashes( sanitize_text_field( $label ) ),
@@ -227,6 +282,14 @@ class PirateForms_Public {
 			/**
 			 ******  Subject field */
 			if ( ! empty( $field ) && ! empty( $label ) ) {
+				$required     = $field === 'req' ? true : false;
+				$wrap_classes = array(
+					'col-xs-12 contact_subject_wrap pirate_forms_three_inputs form_field_wrap',
+				);
+				// If this field was submitted with invalid data
+				if ( isset( $_SESSION[ $error_key ]['contact-subject'] ) ) {
+					$wrap_classes[] = 'error';
+				}
 				$elements[] = array(
 					'front_end'     => true,
 					'placeholder'  => stripslashes( sanitize_text_field( $label ) ),
@@ -262,6 +325,12 @@ class PirateForms_Public {
 			/**
 			 ******  Attachment field */
 			if ( ! empty( $field ) && 'no' !== $field ) {
+				$required     = $field === 'req' ? true : false;
+				$wrap_classes = array( 'col-xs-12 form_field_wrap contact_attachment_wrap' );
+				// If this field was submitted with invalid data
+				if ( isset( $_SESSION[ $error_key ]['contact-attachment'] ) ) {
+					$wrap_classes[] = 'error';
+				}
 				$elements[] = array(
 					'front_end'     => true,
 					'required'     => $field === 'req' ? true : false,
@@ -271,6 +340,37 @@ class PirateForms_Public {
 					'id'           => 'pirate-forms-attachment',
 					'wrap_class'   => isset( $_SESSION[ $error_key ]['contact-attachment'] ) ? 'error' : '',
 				);
+			}
+
+			if ( array_key_exists( 'pirateformsopt_checkbox_field', $pirate_forms_options ) ) {
+				$field = $pirate_forms_options['pirateformsopt_checkbox_field'];
+				$label = $pirate_forms_options['pirateformsopt_label_checkbox'];
+
+				/**
+				 ******  checkbox field */
+				if ( ! empty( $field ) && ! empty( $label ) ) {
+					$required     = $field === 'req' ? true : false;
+					$wrap_classes = array( 'col-xs-12 form_field_wrap contact_checkbox_wrap  ' );
+					// If this field was submitted with invalid data
+					if ( isset( $_SESSION[ $error_key ]['contact-checkbox'] ) ) {
+						$wrap_classes[] = 'error';
+					}
+					$elements[] = array(
+						'required'     => $required,
+						'required_msg' => $pirate_forms_options['pirateformsopt_label_err_no_checkbox'],
+						'type'         => 'checkbox',
+						'class'        => 'form-control',
+						'id'           => 'pirate-forms-contact-checkbox',
+						'wrap'         => array(
+							'type'  => 'div',
+							'class' => implode( ' ', apply_filters( 'pirateform_wrap_classes_checkbox', $wrap_classes ) ),
+						),
+						'value'        => isset( $_REQUEST['pirate-forms-contact-checkbox'] ) ? $_REQUEST['pirate-forms-contact-checkbox'] : '',
+						'options'       => array(
+							'yes'       => stripslashes( $label ),
+						),
+					);
+				}
 			}
 
 			/**
@@ -284,6 +384,11 @@ class PirateForms_Public {
 					'placeholder' => stripslashes( sanitize_text_field( $label ) ),
 					'type'        => 'captcha',
 					'custom'      => array( 'data-sitekey' => $pirateformsopt_recaptcha_sitekey ),
+					'id'          => 'pirate-forms-captcha',
+					'wrap'        => array(
+						'type'  => 'div',
+						'class' => implode( ' ', apply_filters( 'pirateform_wrap_classes_captcha', array( 'col-xs-12 ' . ( $pirate_forms_options['pirateformsopt_recaptcha_field'] !== 'yes' ? 'col-sm-6 ' : '' ) . 'form_field_wrap form_captcha_wrap' ) ) ),
+					),
 				);
 			}
 
@@ -308,6 +413,11 @@ class PirateForms_Public {
 				'front_end' => true,
 				'type'  => 'button',
 				'id'    => 'pirate-forms-contact-submit',
+				'class' => 'pirate-forms-submit-button btn btn-primary',
+				'wrap'  => array(
+					'type'  => 'div',
+					'class' => implode( ' ', apply_filters( 'pirateform_wrap_classes_submit', array( 'col-xs-12 ' . ( ( ! empty( $pirate_forms_options['pirateformsopt_recaptcha_field'] ) && ( $pirate_forms_options['pirateformsopt_recaptcha_field'] !== 'yes' ) ) ? 'col-sm-6 ' : '' ) . 'form_field_wrap contact_submit_wrap' ) ) ),
+				),
 				'value' => $pirateformsopt_label_submit_btn,
 			);
 		}
@@ -405,6 +515,7 @@ class PirateForms_Public {
 		}
 
 		// Start the body of the contact email
+		$private_fields  = array();
 		$body            = array();
 		$body['heading'] = sprintf( __( 'Contact form submission from %s', 'pirate-forms' ), get_bloginfo( 'name' ) . ' (' . site_url() . ')' );
 		$body['body']    = array();
@@ -436,6 +547,9 @@ class PirateForms_Public {
 			$body['body'][ __( 'IP address', 'pirate-forms' ) ] = $contact_ip;
 			$body['body'][ __( 'IP search', 'pirate-forms' ) ]  = "http://whatismyipaddress.com/ip/$contact_ip";
 			$body['magic_tags'] += array( 'ip' => $contact_ip );
+
+			$private_fields[]   = __( 'IP address', 'pirate-forms' );
+			$private_fields[]   = __( 'IP search', 'pirate-forms' );
 		}
 
 		// Sanitize and prepare referrer;
@@ -443,12 +557,16 @@ class PirateForms_Public {
 			$page                                               = sanitize_text_field( $_POST['pirate-forms-contact-referrer'] );
 			$body['body'][ __( 'Came from', 'pirate-forms' ) ]  = $page;
 			$body['magic_tags'] += array( 'referer' => $page );
+
+			$private_fields[]   = __( 'Came from', 'pirate-forms' );
 		}
 
 		// Show the page this contact form was submitted on
 		$permalink                                              = get_permalink( get_the_id() );
 		$body['body'][ __( 'Sent from page', 'pirate-forms' ) ] = $permalink;
 		$body['magic_tags'] += array( 'permalink' => $permalink );
+
+		$private_fields[]   = __( 'Sent from page', 'pirate-forms' );
 
 		// Check the blacklist
 		$blocked = PirateForms_Util::is_blacklisted( $error_key, $pirate_forms_contact_email, $contact_ip );
@@ -458,7 +576,7 @@ class PirateForms_Public {
 		}
 
 		if ( $this->is_spam( $pirate_forms_options, $contact_ip, get_permalink( get_the_id() ), $msg ) ) {
-			$_SESSION[ $error_key ]['honeypot'] = __( 'Form submission failed!', 'pirate-forms' );
+			$_SESSION[ $error_key ]['honeypot'] = __( 'Form submission failed!!', 'pirate-forms' );
 		}
 
 		if ( ! empty( $_SESSION[ $error_key ] ) ) {
@@ -501,15 +619,16 @@ class PirateForms_Public {
 
 			$attachments = $this->get_attachments( $error_key, $pirate_forms_options, $body );
 			if ( is_bool( $attachments ) ) {
+				PirateForms_Util::save_error( $error_key, $nonce_append . '.' . $form_id );
 				return false;
 			}
 
-			$subject = 'Contact on ' . htmlspecialchars_decode( get_bloginfo( 'name' ) );
+			$subject = apply_filters( 'pirate_forms_subject', 'Contact on ' . htmlspecialchars_decode( get_bloginfo( 'name' ) ) );
 			if ( ! empty( $pirate_forms_contact_subject ) ) {
 				$subject = $pirate_forms_contact_subject;
 			}
 
-			$mail_body = ! empty( $pirate_forms_options['pirateformsopt_email_content'] ) ? $pirate_forms_options['pirateformsopt_email_content'] : PirateForms_Util::get_default_email_content( true, $form_id );
+			$mail_body = ! empty( $pirate_forms_options['pirateformsopt_email_content'] ) ? $pirate_forms_options['pirateformsopt_email_content'] : PirateForms_Util::get_default_email_content( true, $form_id, true );
 			$mail_body = PirateForms_Util::replace_magic_tags( $mail_body, $body );
 
 			do_action( 'pirate_forms_before_sending', $pirate_forms_contact_email, $site_recipients, $subject, $mail_body, $headers, $attachments );
@@ -544,6 +663,10 @@ class PirateForms_Public {
 
 				$headers = "From: $site_name <$send_from>\r\nReply-To: $site_name <$send_from>";
 				$subject = $pirate_forms_options['pirateformsopt_label_submit'] . ' - ' . $site_name;
+
+				if ( isset( $pirate_forms_options['pirateformsopt_copy_email'] ) && 'yes' === $pirate_forms_options['pirateformsopt_copy_email'] ) {
+					$confirm_body = $this->append_original_email( $confirm_body, $body, $private_fields );
+				}
 
 				do_action( 'pirate_forms_before_sending_confirm', $pirate_forms_contact_email, $pirate_forms_contact_email, $subject, $confirm_body, $headers );
 				do_action( 'themeisle_log_event', PIRATEFORMS_NAME, sprintf( 'before sending confirm email to = %s, subject = %s, body = %s, headers = %s', $pirate_forms_contact_email, $subject, $confirm_body, $headers ), 'debug', __FILE__, __LINE__ );
@@ -592,6 +715,7 @@ class PirateForms_Public {
 				);
 
 				$redirect_to        = null;
+				$scroll_to          = isset( $_POST['pirate_forms_from_form'] ) ? $_POST['pirate_forms_from_form'] : '';
 
 				/* If a Thank you page is selected, redirect to that page */
 				if ( $pirate_forms_options['pirateformsopt_thank_you_url'] ) {
@@ -602,14 +726,21 @@ class PirateForms_Public {
 					}
 				} elseif ( $is_our_theme ) {
 					// the fragment identifier should always be the last argument, otherwise the thank you message will not show.
+					// the fragment identifier is called pcf here so that the URL can tell us if our theme was recognized.
 					$redirect_to = add_query_arg(
 						array(
 							'done' => 'done',
-							'pcf'  => '1#contact',
+							'pcf'  => "#$scroll_to",
 						), $_SERVER['HTTP_REFERER']
 					);
 				} elseif ( isset( $_SERVER['HTTP_REFERER'] ) ) {
-					$redirect_to = add_query_arg( array( 'done' => 'done' ), $_SERVER['HTTP_REFERER'] );
+					// the fragment identifier is called pf here so that the URL can tell us if this is a not-our theme case.
+					$redirect_to = add_query_arg(
+						array(
+							'done' => 'done',
+							'pf'    => "#$scroll_to",
+						), $_SERVER['HTTP_REFERER']
+					);
 				}
 
 				if ( $redirect_to ) {
@@ -618,6 +749,27 @@ class PirateForms_Public {
 				}
 			}
 		}
+	}
+
+	/**
+	 * Appends the publicly displayed fields to the confirmation email.
+	 *
+	 * @param string $confirm_body The confirmation body.
+	 * @param array  $body The collected body fields.
+	 * @param array  $private_fields The private fields that should not be included from the body fields.
+	 */
+	private function append_original_email( $confirm_body, $body, $private_fields ) {
+		$lines      = array();
+		$lines[]    = '------ Original Email ------';
+
+		foreach ( $body['body'] as $field => $value ) {
+			if ( in_array( $field, $private_fields ) ) {
+				continue;
+			}
+			$lines[] = "$field : $value";
+		}
+
+		return $confirm_body . implode( '<br/>', $lines );
 	}
 
 	/**
@@ -749,6 +901,10 @@ class PirateForms_Public {
 	 * Check with akismet if the message is spam.
 	 */
 	function is_spam( $pirate_forms_options, $ip, $page_url, $msg ) {
+		if ( 'yes' !== PirateForms_Util::get_option( 'pirateformsopt_akismet' ) ) {
+			return false;
+		}
+
 		// check if akismet is installed and key provided
 		$key = get_option( 'wordpress_api_key' );
 		if ( empty( $key ) ) {
@@ -808,13 +964,10 @@ class PirateForms_Public {
 					continue;
 				}
 				/* Validate file type */
-				$file_types_allowed              = implode( '|', apply_filters( 'pirate_forms_allowed_file_types', explode( '|', 'jpg|jpeg|png|gif|pdf|doc|docx|ppt|pptx|odt|avi|ogg|m4a|mov|mp3|mp4|mpg|wav|wmv|xls|xlsx|txt' ) ) );
-				$pirate_forms_file_types_allowed = $file_types_allowed;
-				$pirate_forms_file_types_allowed = trim( $pirate_forms_file_types_allowed, '|' );
-				$pirate_forms_file_types_allowed = '(' . $pirate_forms_file_types_allowed . ')';
-				$pirate_forms_file_types_allowed = '/\.' . $pirate_forms_file_types_allowed . '$/i';
+				$allowed = implode( '|', apply_filters( 'pirate_forms_allowed_file_types', self::$_file_types_allowed ) );
+				$pirate_forms_file_types_allowed = '/\.(' . trim( $allowed, '|' ) . ')$/i';
 				if ( ! preg_match( $pirate_forms_file_types_allowed, $file['name'] ) ) {
-					do_action( 'themeisle_log_event', PIRATEFORMS_NAME, sprintf( 'file invalid: expected %s got %s', $file_types_allowed, $file['name'] ), 'error', __FILE__, __LINE__ );
+					do_action( 'themeisle_log_event', PIRATEFORMS_NAME, sprintf( 'file invalid: expected %s got %s', $allowed, $file['name'] ), 'error', __FILE__, __LINE__ );
 					$_SESSION[ $error_key ]['pirate-forms-upload-failed-type'] = sprintf( __( 'Uploaded file type is not allowed for %s', 'pirate-forms' ), $file['name'] );
 
 					return false;