+++ CHANGELOG.md
@@ -1,4 +1,40 @@
 
+ ### v2.4.0 - 2018-04-02 
+ **Changes:** 
+ * Improves compatibility with various themes.
+* Adds support for two new custom fields.
+* Adds filter for custom classes into form fields.
+* Adds visual/text switch to form wyiwyg editor.
+ 
+ ### v2.3.5 - 2018-03-05 
+ **Changes:** 
+ * Fix characters encoding issue in the subject field.
+* Fix issue with spam label with two forms on the page.
+* Allows zip files to be attached in forms.
+* Adds a filter to dynamically change the subject.
+* Adds options to send a copy of the email to the sender.
+ 
+ ### v2.3.4 - 2018-02-15 
+ **Changes:** 
+ * Added missing Loader.gif file
+* Fixed undefined notice
+* Fix submit button leaving form when ReCaptcha is enabled
+ 
+ ### v2.3.3 - 2018-01-06 
+ **Changes:** 
+ * Fix double reCAPTCHA box bug.
+* Fix custom spam trap alignement error.
+ 
+ ### v2.3.2 - 2017-12-28 
+ **Changes:** 
+ * Fix for tooltip admin behavior.
+ 
+ ### v2.3.1 - 2017-12-28 
+ **Changes:** 
+ * Improves layout and compatibility with various themes.
+* Improves form default email format.
+* Fix issues with various special characters in the magic tag fields.
+ 
  ### v2.3.0 - 2017-11-27 
  **Changes:** 
  * Adds email content wysiwyg editor.